Purchase page

+++ src/pages/CategoriesPage/index.jsx
@@ -1,89 +1,63 @@
-import React, { useEffect, useState } from 'react';
+import React, { useState } from 'react';
 import { useDispatch, useSelector } from "react-redux";
-
-import List from '../../components/List';
-import CategoryRedactor from '../../components/Redactors/CategoryRedactor';
-import {
-    getCategories,
-    addCategory,
-    updateCategory,
-    deleteCategory,
-    selectCategory
-} from "../../redux/category/category.actions";
 import { Button } from 'react-bootstrap';
+
+import {CategoryRedactor, List} from '../../components';
+import {
+    deleteCategory,
+    setCategory
+} from "../../redux/category/category.actions";
 
 import './style.scss'
 
 const CategoriesPage = () => {
+    const dispatch = useDispatch();
     const { categories, isLoading } = useSelector(({ Categories }) => ({
         categories: Categories.list,
         isLoading: Categories.loading
     }))
-    const dispatch = useDispatch();
 
-    const [saveOptions, setSaveOptions] = useState('');
+    const [redactorState, setRedactorState] = useState('');
+    const [showRedactor, setShowRedactor] = useState(false);
 
-    const onSelectCategory = (id) => {
-        console.log(id);
-        if (id === '') {
-            dispatch(selectCategory(id));
-            setShowRedactor(true);
-            setSaveOptions('add');
-        } else {
-            dispatch(selectCategory(id));
-            setShowRedactor(true);
-            setSaveOptions('edit')
-        }
-
-
-
-    }
-
-    useEffect(() => {
-        dispatch(getCategories())
-    }, [dispatch])
-
-
-
-    const onAddCategory = (category) => {
-        dispatch(addCategory(category))
+    const onAddCategory = () => {
+        setRedactorState('add')
+        setShowRedactor(true);
+        dispatch(setCategory(null))
     }
 
     const onEditCategory = (category) => {
-        dispatch(updateCategory(category))
+        setRedactorState('edit')
+        setShowRedactor(true);
+        dispatch(setCategory(category))
     }
 
-    const onDeleteCategory = (id, name) => {
-        if (window.confirm(`Видалити ${name}?`)) {
-            dispatch(deleteCategory(id))
-        }
+    const onDeleteCategory = ({id, name}) => {
+        window.confirm(`Видалити ${name}?`) && dispatch(deleteCategory(id))
     }
-    const [showRedactor, setShowRedactor] = useState(false);
 
     return (
         <div className='page-container'>
             <div className='page-list'>
                 <Button className='list-add-button'
                     variant="primary"
-                    onClick={() => onSelectCategory('')}> Додати +</Button>
+                    onClick={onAddCategory}> Додати +</Button>
                 <List
                     items={categories}
                     isLoading={isLoading}
-                    onSelectItem={onSelectCategory}
+                    onEditItem={onEditCategory}
                     onDeleteItem={onDeleteCategory}
                 />
             </div>
             <div className='page-item'>
                 {showRedactor ? <CategoryRedactor
-                    onAddCategory={onAddCategory}
-                    onEditCategory={onEditCategory}
-                    saveOptions={saveOptions}
+                    redactorState={redactorState}
                 /> :
-                    <div className='page-item-message'>Редагуйте елемет зі списку, або добавте новий</div>}
+                    <div className='page-item-message'>Редагуйте елемет зі списку або добавте новий</div>}
 
             </div>
         </div>
     )
 }
 
-export default CategoriesPage
+export default CategoriesPage;
